organized files, added volume shortcut
