+++ README.md
@@ -87,8 +87,6 @@ cd into $DIR and install requirements
 - document architecture
 - change name
 - add package to package repos (homebrew, apt, etc...)
-- organize codebase
-- create Utils/api.py for spotify API stuff
 
 
 ##### Completed features
